+++ lib/main.dart
@@ -4,19 +4,65 @@ void main() {
   runApp(Home());
 }
 
-class Home extends StatelessWidget {
+class Home extends StatefulWidget {
+  @override
+  _HomeState createState() => _HomeState();
+}
+
+class _HomeState extends State<Home> {
+  int _currentIndex = 0;
+
+  final tabs = [
+    Center(
+      child: Text('Overview'),
+    ),
+    Center(
+      child: Text('Game List'),
+    ),
+    Center(
+      child: Image.network(
+          'https://i.pinimg.com/originals/53/f9/8a/53f98a6b76f60356b2b4c261963377e6.jpg'),
+    )
+  ];
+
   @override
   Widget build(BuildContext context) {
     var title = 'Game List';
     return MaterialApp(
       title: title,
+      debugShowCheckedModeBanner: false,
       home: Scaffold(
         appBar: AppBar(
             title: Row(
           children: [Icon(Icons.gamepad), SizedBox(width: 8), Text(title)],
         )),
-        body: Image.network(
-            'https://i.pinimg.com/originals/53/f9/8a/53f98a6b76f60356b2b4c261963377e6.jpg'),
+        body: tabs[_currentIndex],
+        bottomNavigationBar: BottomNavigationBar(
+          currentIndex: _currentIndex,
+          type: BottomNavigationBarType.fixed,
+          iconSize: 32,
+          selectedFontSize: 12,
+          unselectedFontSize: 12,
+          items: [
+            BottomNavigationBarItem(
+                icon: Icon(Icons.search),
+                label: 'Overview',
+                backgroundColor: Colors.blue),
+            BottomNavigationBarItem(
+                icon: Icon(Icons.gamepad),
+                label: 'Game List',
+                backgroundColor: Colors.blue),
+            BottomNavigationBarItem(
+                icon: Icon(Icons.person),
+                label: 'Profile',
+                backgroundColor: Colors.blue),
+          ],
+          onTap: (index) {
+            setState(() {
+              _currentIndex = index;
+            });
+          },
+        ),
       ),
     );
   }